Add few more kernels to characterize copy/async
The layout of the SMEM buffer that is the destination of the cp.async
affects the number of wavefront required to process the request in both
SMEM and GMEM. The optimal performance is observed when the SMEM range
accessed by the warp is 128B aligned and the range is contiguous.

Need to check how the GMEM layout affects the number of wavefronts
required.